+++ apps/mobile/utils/changelogData.ts
@@ -9,7 +9,23 @@
 /**
  * Raw CHANGELOG.md content embedded at build time.
  */
-export const CHANGELOG_TEXT = `## [1.0.80](https://github.com/alternun-development/alternun/compare/v1.0.79...v1.0.80) (2026-04-16)
+export const CHANGELOG_TEXT = `## [1.0.81](https://github.com/alternun-development/alternun/compare/v1.0.79...v1.0.81) (2026-04-16)
+
+
+### Bug Fixes
+
+* **repo,admin,api:** CHANGELOG, app, build ([6039e90](https://github.com/alternun-development/alternun/commit/6039e9064b183b2969e95183331ff427f3b632d6))
+
+
+
+
+
+## [1.0.81](https://github.com/alternun-development/alternun/compare/v1.0.79...v1.0.81) (2026-04-16)
+
+
+### Bug Fixes
+
+* **repo,admin,api:** CHANGELOG, app, build ([6039e90](https://github.com/alternun-development/alternun/commit/6039e9064b183b2969e95183331ff427f3b632d6))
 
 
 
@@ -17,9 +33,7 @@ export const CHANGELOG_TEXT = `## [1.0.80](https://github.com/alternun-developme
 
 ## [1.0.80](https://github.com/alternun-development/alternun/compare/v1.0.79...v1.0.80) (2026-04-16)
 
-
-
-
+## [1.0.80](https://github.com/alternun-development/alternun/compare/v1.0.79...v1.0.80) (2026-04-16)
 
 ## [1.0.79](https://github.com/alternun-development/alternun/compare/v1.0.78...v1.0.79) (2026-04-16)
 
@@ -1401,4 +1415,4 @@ All notable changes to this project will be documented in this file.
 /**
  * Version extracted from package.json for display purposes.
  */
-export const APP_VERSION = '1.0.80';
+export const APP_VERSION = '1.0.81';